Repair buying.
Use actual timestamps for trade blocks instead of what is
effectively a counter of how many times we had free ships
to perform buys (which might be rarely if we have few ships).

Also, attempt smaller buy transactions.

Also, actually loop through wares when trying to buy wares instead
of missing on the first ware, blocking it, sending a truck to sell
stuff and by the time the truck is free we have unblocked the most
demanded buy ware, miss buying it and send the truck to sell again.
(once again a problem with a small amount of trucks only)